Created NewNoteForm component, which receives props user, team, and addTeamNote from TeamNotes component. Successfully added a new note and got notes in response.
