+++ frontend/src/components/TeamDashboard/TeamNotes/styles/TeamNotes.module.scss
@@ -9,42 +9,3 @@
 h4 {
 	margin: 5px;
 }
-
-.newNoteForm {
-	height: 150px;
-	padding: 0 10px 10px 10px;
-	border-top: 1px solid grey;
-
-	.inputRow {
-		margin: 5px;
-		width: 250px;
-		display: flex;
-		justify-content: space-between;
-
-		input {
-			margin: 0 0 0 5px;
-			width: 170px;
-		}
-	}
-
-	button {
-		margin: 5px 0 5px 0;
-	}
-}
-
-.tagUserMenu {
-	position: relative;
-
-	.tagMenuDropDown {
-		z-index: 100;
-		position: absolute;
-		top: 0;
-		left: 70px;
-		display: flex;
-		flex-direction: column;
-		align-items: flex-start;
-		border: 1px solid grey;
-		border-radius: 5px;
-		background-color: lightgrey;
-	}
-}
